Adds README
The README should give a way to reproduce the results in a manner that
is easy to follow. Therefore, several scripts were created to convert
images to the desired format and resolution, which can be called directly
from the code directory without manually setting any parameters.
Additionally, all paths refering to /data were changed to refer to a local
directory ./data, so everything the scripts do is contained within the main
directory, and no permission issues occur.