Adds info about moving the PSPNet config
The PSPNet config needs to be moved to the semseg/config subdirectory,
in a way that makes it possible for the semseg/tools/train.sh script
to find it there. This commit adds the respective mv command and
changes the directory structure to make that easy for the user to do.